pyproject.toml
[project]
name = "phone-tasker-mcp"
version = "0.1.0"
description = "MCP Server exposing Tasker actions on Android phone as tools"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.4.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
tasker-mcp = "tasker_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tasker_mcp"]
